requirement2.txt
# requirements.txt (with specific versions)

pandas
numpy
SQLAlchemy
psycopg2-binary
python-dotenv